Module 2: Task 4 - version 2

+++ kodilla-intro/src/main/java/Grades.java
@@ -21,7 +21,7 @@ public class Grades {
     }
 
     public double average(int[] grades) {
-        int sum = 0;
+        double sum = 0.0;
         for (int i = 0; i < size; i++) {
             sum = sum + grades[i];
         }
